requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
Flask-Login==0.6.3
python-dotenv==1.0.0
PyPDF2==3.0.1
google-generativeai==0.7.0
email-validator==2.1.0
Werkzeug==3.0.1
